Switch to a more API-like design

+++ src/main/java/nick1st/fancyvideo/BufferToMatrixStack.java
@@ -1,6 +1,7 @@
 package nick1st.fancyvideo;
 
 import com.mojang.blaze3d.matrix.MatrixStack;
+import com.mojang.blaze3d.platform.GlStateManager;
 import com.mojang.blaze3d.systems.RenderSystem;
 import net.minecraft.client.renderer.BufferBuilder;
 import net.minecraft.client.renderer.Tessellator;
@@ -76,7 +77,7 @@ public class BufferToMatrixStack {
         return this;
     }
 
-    public void set(float x, float y, int color) {
+    public void set(int x, int y, int color) {
 
         // Create color
         float r = (color >> 16 & 255) / 255.0F;
